fix: resolve Biome linting violations and improve code quality
- Fix useConsistentMemberAccessibility rule violations by removing explicit public modifiers
- Move UUID regex to top-level scope to fix useTopLevelRegex performance issue
- Disable noSecrets rule to prevent false positives on CSS classes and component names
- Fix daily period date range bug in providers query to cover full day
- Improve provider API validation and error handling with UUID/ULID support
- Enhance error logging for better debugging and monitoring
- Move type definitions from database layer to dedicated types module
- Fix unused variable warning in route handler
